pyproject.toml
[tool.poetry]
name = "thriftease-api"
version = "0.1.0"
description = ""
authors = ["enaielei <nommel.isanar.lavapie.amolat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
graphene-django = "^3.1.5"
django-cors-headers = "^4.3.1"
django-filter = "^23.5"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.7"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.7"}
mypy = "^1.7.1"
polib = "^1.2.0"
types-polib = "^1.2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"